store_order_in_db

+++ app/Http/Controllers/CheckoutController.php
@@ -2,11 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use DateTime;
+use App\Order;
 use Gloudemans\Shoppingcart\Facades\Cart;
 use Stripe\Stripe;
 use Stripe\PaymentIntent;
 use Illuminate\Http\Request;
 use Illuminate\Support\Arr;
+use Illuminate\Support\Facades\Session;
 
 class CheckoutController extends Controller
 {
@@ -59,11 +62,41 @@ class CheckoutController extends Controller
      */
     public function store(Request $request)
     {
-        Cart::destroy();
-
         $data = $request->json()->all();
 
-        return $data['paymentIntent'];
+        $order = new Order();
+
+        $order->payment_intent_id = $data['paymentIntent']['id'];
+        $order->amount = $data['paymentIntent']['amount'];
+
+        $order->payment_created_at = (new DateTime())->setTimestamp($data['paymentIntent']['created'])->format('Y-m-d H:i:s');
+
+        $products = [];
+        $i = 0;
+        
+        foreach (Cart::content() as $product) {
+            $products['product_' . $i][] = $product->model->title;
+            $products['product_' . $i][] = $product->model->price;
+            $products['product_' . $i][] = $product->qty;
+            $i++;
+        }
+
+        $order->products = serialize($products);
+        $order->user_id = 15;
+        $order->save();
+
+        if ($data['paymentIntent']['status'] === 'succeeded') {
+            Cart::destroy();
+            Session::flash('success', 'Votre commande a été traitée avec succès.');
+            return response()->json(['success' => 'Payment Intent Succeeded']);
+
+        } else {
+            return response()->json(['error' => 'Payment Intent Not Succeeded']);
+        }
+    }
+
+    public function thankyou(){
+        return Session::has('success') ? view('checkout.thankYou') : redirect()->route('products.index');
     }
 
     /**
